ajuste no loop verificaçao sem net

+++ lib/services/heroes/store/heroes_store.dart
@@ -45,8 +45,10 @@ class HeroesStore {
                 categoryId: data.categoryId,
                 name: data.name,
               );
-              await registerHeroes(requestData);
-              await requestBaseDao.deleteRequest(data.created!);
+              final result = await registerHeroes(requestData);
+              if (result.isRight()) {
+                await requestBaseDao.deleteRequest(data.created!);
+              }
               break;
             case "PUT":
               final requestData = RequestModel(
@@ -55,8 +57,10 @@ class HeroesStore {
                 name: data.name,
                 id: data.id,
               );
-              await updateHeroes(requestData);
-              await requestBaseDao.deleteRequest(data.created!);
+              final result = await updateHeroes(requestData);
+              if (result.isRight()) {
+                await requestBaseDao.deleteRequest(data.created!);
+              }
               break;
             case "DELETE":
               final requestData = RequestModel(
@@ -65,8 +69,10 @@ class HeroesStore {
                 name: data.name,
                 id: data.id,
               );
-              await deleteHeroes(requestData);
-              await requestBaseDao.deleteRequest(data.created!);
+              final result = await deleteHeroes(requestData);
+              if (result.isRight()) {
+                await requestBaseDao.deleteRequest(data.created!);
+              }
               break;
             default:
           }
